💄 slight refactoring & adding some vague unit tests...
